Add update handler

+++ postgres/11-create-table.sql
@@ -3,7 +3,7 @@
 CREATE TABLE IF NOT EXISTS tb_persons(
     id serial PRIMARY KEY,
     name varchar(128),
-    age bigint,
+    age int,
     address varchar(256),
     work varchar(256)
 );
